(issue 39) : move tests from api module to test_api_larray new module + print message when models are reset if test module called with option -O (debug mode)
