+++ setup.cfg
@@ -3,6 +3,6 @@ test=pytest
 
 [tool:pytest]
 testpaths = larray_editor
-addopts = -v --doctest-modules
+addopts = -v --doctest-modules --ignore=tests/test_api_larray.py
 #--maxfail=1 --cov (requires pytest-cov) --pep8 (requires pytest-pep8)
 #pep8maxlinelength = 119
